# openapi.json
{"openapi":"3.1.0","info":{"title":"SAIL Starter Kit","description":"SAIL Starter Kit API with MongoDB Backend","version":"1.0.0"},"paths":{"/api/auth/login":{"post":{"tags":["auth"],"summary":"Login","description":"Login with email and password. Returns JWT token on success.","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/signup":{"post":{"tags":["auth"],"summary":"Signup","description":"Create a new user account. Sends OTP for email verification.","operationId":"signup_api_auth_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Signup Api Auth Signup Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/verify-otp":{"post":{"tags":["auth"],"summary":"Verify Otp","description":"Verify OTP and activate account. Returns JWT token on success.","operationId":"verify_otp_api_auth_verify_otp_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyOTPRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/forgot-password":{"post":{"tags":["auth"],"summary":"Forgot Password","description":"Request password reset. Sends OTP to email.","operationId":"forgot_password_api_auth_forgot_password_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Forgot Password Api Auth Forgot Password Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/reset-password":{"post":{"tags":["auth"],"summary":"Reset Password","description":"Reset password using OTP.","operationId":"reset_password_api_auth_reset_password_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Reset Password Api Auth Reset Password Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/me":{"get":{"tags":["auth"],"summary":"Get Me","description":"Get current user info.","operationId":"get_me_api_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/app__modules__auth__schemas__UserResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/auth/resend-otp":{"post":{"tags":["auth"],"summary":"Resend Otp","description":"Resend OTP for email verification or password reset.","operationId":"resend_otp_api_auth_resend_otp_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Resend Otp Api Auth Resend Otp Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/request-login-otp":{"post":{"tags":["auth"],"summary":"Request Login Otp","description":"Request OTP for passwordless login. Sends OTP to email.","operationId":"request_login_otp_api_auth_request_login_otp_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestLoginOTPRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Request Login Otp Api Auth Request Login Otp Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/verify-login-otp":{"post":{"tags":["auth"],"summary":"Verify Login Otp","description":"Verify OTP for passwordless login. Returns JWT token on success.","operationId":"verify_login_otp_api_auth_verify_login_otp_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyLoginOTPRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/health/":{"get":{"tags":["health"],"summary":"Health Check","description":"Basic health check endpoint.","operationId":"health_check_api_health__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/health/db":{"get":{"tags":["health"],"summary":"Db Health Check","description":"Check MongoDB connectivity.","operationId":"db_health_check_api_health_db_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/health/stats":{"get":{"tags":["health"],"summary":"Db Stats","description":"Get database statistics.","operationId":"db_stats_api_health_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/admin/domains":{"get":{"tags":["admin"],"summary":"List Domains","description":"List all enterprise domains.","operationId":"list_domains_api_admin_domains_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/DomainResponse"},"type":"array","title":"Response List Domains Api Admin Domains Get"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["admin"],"summary":"Create Domain","description":"Create a new enterprise domain.","operationId":"create_domain_api_admin_domains_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDomainRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DomainResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/admin/domains/{domain_uuid}/modules":{"put":{"tags":["admin"],"summary":"Update Domain Modules","description":"Update enabled modules for a domain.","operationId":"update_domain_modules_api_admin_domains__domain_uuid__modules_put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"domain_uuid","in":"path","required":true,"schema":{"type":"string","title":"Domain Uuid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDomainModulesRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DomainResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/admin/stats":{"get":{"tags":["admin"],"summary":"Get Admin Stats","description":"Get admin dashboard statistics.","operationId":"get_admin_stats_api_admin_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminStatsResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/admin/users":{"get":{"tags":["admin"],"summary":"List Users","description":"List all users with filtering and pagination.","operationId":"list_users_api_admin_users_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Search by email or name","title":"Search"},"description":"Search by email or name"},{"name":"role","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by role","title":"Role"},"description":"Filter by role"},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by status","title":"Status"},"description":"Filter by status"},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Page number","default":1,"title":"Page"},"description":"Page number"},{"name":"per_page","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Items per page","default":20,"title":"Per Page"},"description":"Items per page"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["admin"],"summary":"Create User","description":"Create a new user (bypass normal signup flow).","operationId":"create_user_api_admin_users_post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/app__modules__admin__schemas__UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/admin/users/{user_id}":{"get":{"tags":["admin"],"summary":"Get User","description":"Get a single user by ID.","operationId":"get_user_api_admin_users__user_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/app__modules__admin__schemas__UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["admin"],"summary":"Update User","description":"Update a user.","operationId":"update_user_api_admin_users__user_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/app__modules__admin__schemas__UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["admin"],"summary":"Delete User","description":"Permanently delete a user.","operationId":"delete_user_api_admin_users__user_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/account":{"get":{"tags":["account"],"summary":"Get Account","description":"Get current user account profile.","operationId":"get_account_api_account_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}}},"security":[{"HTTPBearer":[]}]},"put":{"tags":["account"],"summary":"Update Account","description":"Update current user account profile.","operationId":"update_account_api_account_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountUpdateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/account/change-password":{"post":{"tags":["account"],"summary":"Change Password","description":"Change current user password.","operationId":"change_password_api_account_change_password_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordChangeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/quiz/complete":{"post":{"tags":["quiz"],"summary":"Complete Quiz","description":"Record quiz completion and send celebration email if perfect score.","operationId":"complete_quiz_api_quiz_complete_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizCompleteRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/quiz/certificate":{"post":{"tags":["quiz"],"summary":"Issue Certificate","description":"Issue a certificate of completion.","operationId":"issue_certificate_api_quiz_certificate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CertificateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/":{"get":{"summary":"Root","description":"Root endpoint.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AccountResponse":{"properties":{"uuid":{"type":"string","title":"Uuid"},"email":{"type":"string","title":"Email"},"first_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"First Name"},"last_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Name"},"department":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Department"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number"},"role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Role"},"status":{"type":"string","title":"Status"}},"type":"object","required":["uuid","email","status"],"title":"AccountResponse","description":"Account profile response."},"AccountUpdateRequest":{"properties":{"first_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"First Name"},"last_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Name"},"department":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Department"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number"}},"type":"object","title":"AccountUpdateRequest","description":"Request to update account profile."},"AdminStatsResponse":{"properties":{"total_users":{"type":"integer","title":"Total Users"},"active_users":{"type":"integer","title":"Active Users"},"pending_users":{"type":"integer","title":"Pending Users"},"suspended_users":{"type":"integer","title":"Suspended Users"},"super_admins":{"type":"integer","title":"Super Admins"},"by_status":{"additionalProperties":{"type":"integer"},"type":"object","title":"By Status"}},"type":"object","required":["total_users","active_users","pending_users","suspended_users","super_admins","by_status"],"title":"AdminStatsResponse","description":"Response schema for admin dashboard stats."},"CertificateRequest":{"properties":{"course_id":{"type":"string","title":"Course Id"},"course_title":{"type":"string","title":"Course Title"},"user_name":{"type":"string","title":"User Name"},"date":{"type":"string","title":"Date"},"supervisor_email":{"type":"string","format":"email","title":"Supervisor Email"}},"type":"object","required":["course_id","course_title","user_name","date","supervisor_email"],"title":"CertificateRequest"},"CreateDomainRequest":{"properties":{"domain":{"type":"string","title":"Domain"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"enabled_modules":{"items":{"type":"string"},"type":"array","title":"Enabled Modules","default":[]}},"type":"object","required":["domain"],"title":"CreateDomainRequest"},"CreateUserRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","minLength":8,"title":"Password"},"first_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"First Name"},"last_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Name"},"department":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Department"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number"},"role":{"anyOf":[{"type":"string","enum":["super_admin","admin","editor","viewer"]},{"type":"null"}],"title":"Role"},"status":{"type":"string","enum":["pending","active","suspended"],"title":"Status","default":"active"}},"type":"object","required":["email","password"],"title":"CreateUserRequest","description":"Request schema for creating a new user."},"DomainResponse":{"properties":{"uuid":{"type":"string","title":"Uuid"},"domain":{"type":"string","title":"Domain"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"enabled_modules":{"items":{"type":"string"},"type":"array","title":"Enabled Modules"},"is_active":{"type":"boolean","title":"Is Active"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["uuid","domain","name","enabled_modules","is_active","created_at","updated_at"],"title":"DomainResponse"},"ForgotPasswordRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"}},"type":"object","required":["email"],"title":"ForgotPasswordRequest","description":"Forgot password request model."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest","description":"Login request model."},"PasswordChangeRequest":{"properties":{"current_password":{"type":"string","title":"Current Password"},"new_password":{"type":"string","title":"New Password"}},"type":"object","required":["current_password","new_password"],"title":"PasswordChangeRequest","description":"Request to change password."},"QuizCompleteRequest":{"properties":{"quiz_title":{"type":"string","title":"Quiz Title"},"score":{"type":"integer","title":"Score"},"total_questions":{"type":"integer","title":"Total Questions"}},"type":"object","required":["quiz_title","score","total_questions"],"title":"QuizCompleteRequest"},"RequestLoginOTPRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"}},"type":"object","required":["email"],"title":"RequestLoginOTPRequest","description":"Request OTP for passwordless login."},"ResetPasswordRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"otp":{"type":"string","title":"Otp"},"new_password":{"type":"string","minLength":8,"title":"New Password"}},"type":"object","required":["email","otp","new_password"],"title":"ResetPasswordRequest","description":"Reset password request model."},"SignupRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","minLength":8,"title":"Password"},"first_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"First Name"},"last_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Name"}},"type":"object","required":["email","password"],"title":"SignupRequest","description":"Signup request model."},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"expires_at":{"type":"string","format":"date-time","title":"Expires At"},"user":{"additionalProperties":true,"type":"object","title":"User"}},"type":"object","required":["access_token","expires_at","user"],"title":"TokenResponse","description":"Response returned after successful login."},"UpdateDomainModulesRequest":{"properties":{"enabled_modules":{"items":{"type":"string"},"type":"array","title":"Enabled Modules"}},"type":"object","required":["enabled_modules"],"title":"UpdateDomainModulesRequest"},"UpdateUserRequest":{"properties":{"first_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"First Name"},"last_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Name"},"department":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Department"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number"},"role":{"anyOf":[{"type":"string","enum":["super_admin","admin","editor","viewer"]},{"type":"null"}],"title":"Role"},"status":{"anyOf":[{"type":"string","enum":["pending","active","suspended"]},{"type":"null"}],"title":"Status"},"password":{"anyOf":[{"type":"string","minLength":8},{"type":"null"}],"title":"Password"}},"type":"object","title":"UpdateUserRequest","description":"Request schema for updating a user."},"UserListResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/app__modules__admin__schemas__UserResponse"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"},"page":{"type":"integer","title":"Page"},"per_page":{"type":"integer","title":"Per Page"},"pages":{"type":"integer","title":"Pages"}},"type":"object","required":["items","total","page","per_page","pages"],"title":"UserListResponse","description":"Response schema for user list with pagination."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VerifyLoginOTPRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"otp":{"type":"string","maxLength":6,"minLength":6,"title":"Otp"}},"type":"object","required":["email","otp"],"title":"VerifyLoginOTPRequest","description":"Verify OTP for passwordless login."},"VerifyOTPRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"otp":{"type":"string","title":"Otp"}},"type":"object","required":["email","otp"],"title":"VerifyOTPRequest","description":"OTP verification request model."},"app__modules__admin__schemas__UserResponse":{"properties":{"uuid":{"type":"string","title":"Uuid"},"email":{"type":"string","title":"Email"},"first_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"First Name"},"last_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Name"},"department":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Department"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number"},"role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Role"},"status":{"type":"string","title":"Status"},"created_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"},"last_login_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Login At"}},"type":"object","required":["uuid","email","status"],"title":"UserResponse","description":"Response schema for a user."},"app__modules__auth__schemas__UserResponse":{"properties":{"uuid":{"type":"string","title":"Uuid"},"email":{"type":"string","title":"Email"},"role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Role"},"status":{"type":"string","title":"Status"},"first_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"First Name"},"last_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Name"},"department":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Department"}},"type":"object","required":["uuid","email","status"],"title":"UserResponse","description":"User response model."}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}